requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
joblib
google-cloud-storage
google-cloud-bigquery
